[Docs] Handling routes with no handlers
