+++ docs/docs/cookbook/404-page.md
@@ -7,6 +7,8 @@ title: 404 Page
 Here's a way to implement custom 404 pages.
 
 ```typescript
+import { Get, HttpResponseNotFound, HttpResponseOK } from '@foal/core';
+
 class ViewController {
   @Get('/home')
   home() {
@@ -19,7 +21,22 @@ class AppController {
 
   @Get('*')
   notFound() {
-    return new HttpResponseNotFound('The page your are looking for does not exist');
+    return new HttpResponseNotFound('The page you are looking for does not exist.');
+  }
+}
+```
+
+In case you want to intercept all HTTP verbs (POST, PUT, etc), you can also use the `@All` decorator for this.
+
+```typescript
+import { All, HttpResponseNotFound } from '@foal/core';
+
+class AppController {
+  subControllers = [ ViewController ];
+
+  @All('*')
+  notFound() {
+    return new HttpResponseNotFound('The route you are looking for does not exist.');
   }
 }
 ```
